+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - TODO
+
+### Added
+- `Slang` to `glsl` shader compilation is now allowed.
+- `CompileSlangToGlsl` method was added to the `ShaderCompiler`.
+
 ## [0.1.0] - 2026-03-04
 
 ### Added